Cleanup resource picker component
